[simple-cipher] Combine the two test classes into a single suite
