--- exercises/practice/simple-cipher/.docs/instructions.append.md
@@ -1,8 +1,0 @@
-# Instructions append
-
-This exercise has two test classes: SimpleCipherStepOneTest.java and SimpleCipherStepTwoSubsitutionTest.java.
-
-To run only one test class (e.g. in this example SimpleCipherStepOneTest.java):
-```
-$ gradle test --tests *StepOne*
-```
