--- exercises/practice/simple-cipher/src/test/java/SimpleCipherStepOneTest.java
@@ -1,48 +1,0 @@
-import org.junit.Before;
-import org.junit.Ignore;
-import org.junit.Test;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-/**
- * Step 1: Make a simple shift cipher
- */
-public class SimpleCipherStepOneTest {
-    private Cipher cipherWithDefaultKey;
-
-    @Before
-    public void setup() {
-        cipherWithDefaultKey = new Cipher();
-    }
-
-    /**
-     * Here we take advantage of the fact that plaintext of "aaa..." doesn't output the key. This is a critical problem
-     * with shift ciphers, some characters will always output the key verbatim.
-     */
-    @Test
-    public void cipherCanEncode() {
-        String plainText = "aaaaaaaaaa";
-        String cipherText = cipherWithDefaultKey.getKey().substring(0, 10);
-        assertThat(cipherWithDefaultKey.encode(plainText)).isEqualTo(cipherText);
-    }
-
-    @Ignore("Remove to run test")
-    @Test
-    public void cipherCanDecode() {
-        String cipherText = "aaaaaaaaaa";
-        assertThat(cipherWithDefaultKey.decode(cipherWithDefaultKey.getKey().substring(0, 10))).isEqualTo(cipherText);
-    }
-
-    @Ignore("Remove to run test")
-    @Test
-    public void cipherIsReversible() {
-        String plainText = "abcdefghij";
-        assertThat(cipherWithDefaultKey.decode(cipherWithDefaultKey.encode(plainText))).isEqualTo(plainText);
-    }
-
-    @Ignore("Remove to run test")
-    @Test
-    public void keyIsLowercaseLetters() {
-        assertThat(cipherWithDefaultKey.getKey()).matches("^[a-z]+$");
-    }
-}
